Realign JobsApp partial files with their names
Move BuildDataTableUpdates from JobsApp.Data.cs to JobsApp.DataTable.cs
Move CanRerun from JobsApp.DataTable.cs to JobsApp.Helpers.cs
Move JobRerunEligibilityTests to Apps/Jobs subdirectory

Plan 00476
